ai-sports-mcp/server/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# AI/ML
openai==1.58.1
langchain==0.3.12
langchain-openai==0.2.12
langchain-community==0.3.12
tiktoken==0.8.0

# Vector DB
chromadb==0.5.23
sentence-transformers==3.3.1

# Auth
firebase-admin==6.6.0
python-jose[cryptography]==3.3.0
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
tenacity==9.0.0
redis==5.2.1
celery==5.4.0

# PDF Processing
pypdf==5.1.0
pdfplumber==0.11.4
langchain-text-splitters==0.3.3

# Monitoring
loguru==0.7.3
prometheus-client==0.21.0

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
